Extract `TrikotViewModelDeclarative` to its own file
