+++ trikot-viewmodels-declarative-flow/swift/swiftui/ImageResource/VMDImageProvider.swift
@@ -9,39 +9,3 @@ public struct DefaultImageProvider: VMDImageProvider {
     public init() { }
     public func imageForResource(imageResource: VMDImageResource) -> Image? { nil }
 }
-
-public class TrikotViewModelDeclarative {
-
-    public static let shared: TrikotViewModelDeclarative = TrikotViewModelDeclarative()
-
-    var imageProvider: VMDImageProvider {
-        guard let internalImageProvider = internalImageProvider else {
-            fatalError("TrikotViewModelDeclarative must be initialized before use")
-        }
-
-        return internalImageProvider
-    }
-
-    var spanStyleProvider: VMDSpanStyleProvider {
-        guard let internalSpanStyleProvider = internalSpanStyleProvider else {
-            fatalError("TrikotViewModelDeclarative must be initialized before use")
-        }
-
-        return internalSpanStyleProvider
-    }
-
-    private var internalImageProvider: VMDImageProvider?
-
-    private var internalSpanStyleProvider: VMDSpanStyleProvider?
-
-    private init() {
-    }
-
-    public func initialize(
-        imageProvider: VMDImageProvider = DefaultImageProvider(),
-        spanStyleProvider: VMDSpanStyleProvider = DefaultSpanStyleProvider()
-    ) {
-        internalImageProvider = imageProvider
-        internalSpanStyleProvider = spanStyleProvider
-    }
-}
